commit initial changes
